refactor: SearchDialog filter 컴포넌트 분리
